LayoutTests/fast/exclusions/css-exclusions-disabled-expected.txt

Test that parsing of css exclusions related properties can be disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testInlineStyle("-webkit-wrap-flow", "both") is null
PASS testComputedStyle("-webkit-wrap-flow", "both") is "auto"
PASS testInlineStyle("-webkit-wrap-through", "none") is null
PASS testComputedStyle("-webkit-wrap-through", "none") is "wrap"
PASS successfullyParsed is true

TEST COMPLETE
